Copy desc0, desc and record descriptors to interpreter
See also #43.

In the direction host->interpreter, we need to have proper descriptors
instead of HNF indirections, because the interpreted code may depend on
the structure of a node of which the descriptor has been stripped. This
happens for instance when a repl_r_args is used for a descriptor which
does not exist in the code (and is hence stripped).

This problem does not exist in the other direction, because the host
cannot depend on the structure of nodes in the interpreter, so it does
not necessitate copying descriptors in the other direction as well.

Also note that we do not include code addresses in the descriptors.
Since we do not copy curried nodes, this is not needed.